requirements.txt
astor==0.8.1
certifi==2022.12.7
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1666183775483/work
charset-normalizer==3.1.0
cloudpickle==2.2.1
colorama==0.4.6
contextlib2==21.6.0
cycler==0.11.0
deeprobust==0.2.8
filelock==3.10.7
fonttools==4.38.0
gensim==3.8.3
googledrivedownloader==0.4
idna==3.4
imageio==2.27.0
importlib-metadata==4.13.0
isodate==0.6.1
Jinja2==3.1.2
joblib==1.2.0
json-tricks==3.16.1
kiwisolver==1.4.4
littleutils==0.2.2
llvmlite==0.39.1
MarkupSafe==2.1.2
matplotlib==3.5.3
networkx==2.6.3
nni==2.10
numba==0.56.4
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1649806299270/work
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
ogb==1.3.6
outdated==0.2.2
packaging==23.0
pandas==1.3.5
Pillow==9.4.0
prettytable==3.6.0
protobuf==3.20.3
psutil==5.9.5
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
pyparsing==3.0.9
python-dateutil==2.8.2
PythonWebHDFS==0.2.3
pytz==2023.3
PyWavelets==1.3.0
PyYAML==6.0
rdflib==6.3.2
requests==2.28.2
responses==0.23.1
schema==0.7.5
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
simplejson==3.18.4
six==1.16.0
smart-open==6.3.0
tensorboardX==2.6
texttable==1.6.7
threadpoolctl==3.1.0
tifffile==2021.11.2
torch==1.12.1.post201
torch-cluster==1.6.1
torch-geometric==2.3.1
torch-scatter==2.0.9
torch-sparse==0.6.15
torchvision==0.14.1
tqdm==4.65.0
typeguard==3.0.2
types-PyYAML==6.0.12.9
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1678559861143/work
urllib3==1.26.15
wcwidth==0.2.6
websockets==10.4
yacs==0.1.8
zipp==3.15.0
